move model dir
